requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
faiss-cpu==1.7.4
fastapi==0.109.2
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
idna==3.6
numpy==1.26.4
protobuf==4.25.3
pydantic==2.6.1
pydantic_core==2.16.2
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
